libpayload: Fix type issues
There were a number of type issues in libpayload that sneaked in
with 903f8e0.

- size_t and ssize_t were conflicting with gcc builtins
- some stdint types were used in libpayload but not defined
  in our stdint.h

With this patch it's possible to compile libpayload with the
reference toolchain again.

Reviewed-on: http://review.coreboot.org/2903
Tested-by: build bot (Jenkins)

+++ payloads/libpayload/include/x86/arch/types.h
@@ -53,8 +53,8 @@ typedef signed long long s64;
 typedef long time_t;
 typedef long suseconds_t;
 
-typedef unsigned int size_t;
-typedef int ssize_t;
+typedef unsigned long size_t;
+typedef long ssize_t;
 
 #ifndef NULL
 #define NULL ((void *)0)
